ai.kroaddy.site/services/mlservice/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
icecream>=2.1.3
openpyxl>=3.1.0
xlrd>=2.0.1
googlemaps>=4.10.0
shapely>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0


folium>=0.12.0
matplotlib>=3.7.0
seaborn>=0.12.0

# NLP 패키지
nltk>=3.8.0
wordcloud>=1.9.0
sentencepiece>=0.1.99
konlpy>=0.6.0
opencv-python>=4.8.0
Pillow>=10.0.0
jpype1>=1.4.0